feat: Memo TripItem according to timeserie rev
part of https://github.com/cozy/coachCO2/issues/26
Force not to render the tripItem if the rev does not change. For the record, we can't change the data without using save.client, so the rev is bound to change.
This greatly improves performance when modifying a trip and removes the freeze that was present after this operation.