feat: Aggregate only necessary fields
The `computeAggregatedTimeseries` used to return the whole section,
including arrays with all the speeds, timestamps, duration, etc.
We change this behaviour to only return the necessary aggregated fields.
We expect to eventually store the resulted `aggregation` object in
database, in order to request only this minimal object. This will be
useful to:
- Make requests less heavy and more performant, so the display will be
speed up
- Allow to compute efficient aggregations, which was impossible to do in the
app with so many data to handle.

feat: Add calories, modes and purpose in aggregation