feat: Migrate timeseries without aggregation
We now want to store an `aggregation` object in the
`io.cozy.timeseries.geojson` doctype, so it can be directly queried from
the app without having to retrieve the full geojson object, which can
very heavy, with most of the information being useless.
This service queries all the timeseries without the `aggregation` object
thanks to a partial filer and store the date of the last processed
timeserie, so it can be run incrementally.

refactor: Follow naming conventions