feat: Add service to send monthly CO2 emissions to DACC
This service will be launched every month. For each month, it will
aggregate the CO2 emitted by all the trips for this month, and the send
the result to the DACC.
Eventually, this will give the user the possibility to compare herself
with others contributors.

Note the service can only 12 contributions for each account for a single
execution, to prevent service timeout. If there are more months to
compute, the service relaunch itself.