fix: Upgrade cozy-client to fix query with partial index
When an index with a partial filter was created and an other index
already existed with the same selector, it created a conflict and the
index with the partial filter wasn't created and the one without was
actually used.
We met this situation for the migration service, which was stucked with
th 1000 first docs, because the query was returning all the docs, as the
partial index was not evaluated.