fix: Index on endDate rather than startDate
We have side-effects on instances already having an index on
`startDate` but without a partial index. In this situation, CouchDB will
fallback on this index, and won't apply the partial filter, leading to
unexpected results.
By the time we fix this issue in cozy-client, we use `endDate` as index
attribute.