fix: Compute duration based on aggregation duration
There are 2 queries retrieving trips on the home page:

1. buildAggregatedTimeseriesQueryByAccountId
2. buildOneYearOldTimeseriesWithAggregationByAccountId

The first one requires both startDate and endDate attributes and is
limited to 50 results, while the second one only require startDate but
involves the whole past year.

When the second query is run, there are trips in the store without the
endDate attribute. In the meantime, the Trips component will see those
new trips and display them ; but as the endDate attribute is required,
this was causing an error and a blank page.

This behaviour raises questions about the store management that should
be discussed. In the meantime, this fixes the app issue by removing the
need of endDate.