feat: Add service to set automatic purposes
This service is used to detect similar/recurring trips compared to a
trip with a manual purpose.
The detected recurring trips are then updated with this manual purpose
and are magically displayed in realtime to the user.

Note the recurring trips detection is very naïve for now: we simply
detect trips with the same (or reverse) start/end displayed place.
This is a first step and eventually will be improved later.

feat: Add helpers

feat: Consider automatic purpose