feat: Exclude non-recurring trips
When the user explicitely define a trip as non recurring, it should be
excluded of the considered trips for automatic purposes.
Likewise, any trip with an explicit recurring flag is seen as recurring
and can have an automatic purpose. Even if there is no other similar
trips found, there might be in the future and this flag is thus used to
know if a trip should be considered or not.