feat: Start recurring service after trips creation
When a connector imports trips, the service should be started, so newly
created trips might have an automatic purpose.

Note the behaviour is not the same depending on if the service is run
from the app or the trigger. In case of the app, we are looking for
similar existing trips to set them the new manual purpose. In case of a
trigger, we are searching for any trips without purpose and find similar
trips with manual purpose.