feat(Queries): Change approach to get bike commute timeseries
Because of the way couch works, using elemMatch means that we don't rely on the index and so we browse the whole database. This is suboptimal. But it is a necessary discriminant.
Another point, we delete the date selector here. This made sense if we had optimal queries, and then execute a query by year, when the year changes on the front.
Finally, to know the dates to select, we have to go through all the trips anyway.
With Paul we decided to make only one query and then to sort the data on the front.