feat(geolocation): Use device name with date as login when similar account already exists
When we want to create a new openpath account, and we find that there
are already accounts with the device name as login, instead of
creating a new account with the device name, we create an account
with '<device name> créé le <date>' like 'iPhone créé le 03/10/2023'.

So first account created when enabling geolocation will be 'iPhone'
and then it will be 'iPhone créé le XX/XX/XXXX'.