feat: Manage geolocation permissions
Show a modal to prepare the user to accept permissions.

Show a modal which redirect to OS native settings when the user refused permissions.

> Full explanation about special cases below <

plugin = react-native-background-geolocation

On iOS, the plugin manages permission requests.
On Android, we manage permission requests except for the background, which is managed by the plugin.

You can't do both in the same way because :
- On iOS, if you manage requests, the background request will only be made when the background is actually triggered (i.e. after 200m of travel). I preferred to have all permissions requested at the beginning to avoid annoying the user afterwards. I imagine that the plugin has a native trick for triggering this request for permissions straight away.
- On Android, if the plugin manages requests, you can't display the refused permissions modal properly because you have to make a permission request to find out if a permission is refused. So we'd have the native popups before the user preparation modal.