feat: Filter recurring loop trips on distance
When trying to find recurring timeseries with same purpose, 
 we rely on the coordinates, as we consider that trips with
same start / end points have quite often the same purpose.
However, when the start / end points are the same, this assumption seems less relevant: typically trips from and to the home address can have multiple purposes.
That is why we introduce a special case for "loop trip":  we filter out trips with a too high difference in distance  to only set  same purposes to roughly-same-distance loop trips.
This is obviously not perfect, and should probably be eventually improved, but might be a good enough quick win for now.