fix: Get correct title when contacts are associated to trips
The recurrence service is able to detect when a trip start or end should
be associated with a contact's address, because of their coordinates.

However, this was not correctly working, because the
`makeAggregationTitle` was expecting hydrated timeseries, i.e. including
complete contacts in their docs.
ATM this must be done manually in node.
See https://github.com/cozy/cozy-client/issues/493