fix: Correctly handle realtime for contact creation
When creating a new contact, a Cozy Contact service is triggered with a
5 sec debounce. This service updates the contact and create a new _rev,
which can cause conflicts. And when conflicts arise, the contact is
never actually linked to the trip in database, making the feature
pointless. Even worse, the user was never informed about it, the error
being silent.

Fortunately, the Cozy realtime is there to dynamically updates the contact
with the new data.
Unfortunately, this was not properly handled: the contact itself was
never queried, and, once queried, never set in the state.

The conflicts should now be avoided, making the world a better place.