fix: Avoid blinking page on load
On coachco2 loading, we can see page switching very quickly between
InstallApp and Welcome page because by default we displayed
InstallApp page, and after a request to the flagship app we decided
to stay on this page or change to Welcome page.

Now we wait to be sure of the value of isGeolocationTrackingAvailable
and isGeolocationTrackingEnabled before displaying something.