fix: Remove redirect
Since the navLink "Analysis" was targeting `/analysis` and that
`/analysis` is redirecting to `/analysis/mode`, we had a visual
glitch because we triggered a full redirection and then a full render
of the components.

Here we do 2 things:
- The first one is to make the NavLink directly to `/analysis/mode` like
that we avoid the redirection and the visual glitch
- But with that, we need to handle manually the fact that the
NavLink "Analysis" is active or not (since by default React Router set
it active only if it's a children. Since `/analysis/purposes` is not a
children of `analysis/mode` when we selected "Purpose" the "Analysis"
was not active. So we check the pathName and we set the right className
by importing them from UI.