fix: No desynchronization between isAccountLoading & selectedAccount
With the previous implementation (you can check it out in the git
history),we had a desynchronization between the isAccountLoading
state and the selectedAccount state.

This is because the selectedAccount was only triggered in an other
useEffect because the account was different than the selectedAccount.

In order to fix this issue, we create a ref that tell us if we are
at the "mount time" aka if the accountRef.current is null. If yes,
then we can set directly the selectedAccount  from the account.

If not, then we wait for the account to be different than the
selectedAccount and then we set it.