fix: Reserve the height for the graph
Since we render the CO2EmissionsChart only when we have an account & a
timeseries, we are sur to display it when we render it.

So this is a quick fix to avoid visual glitch by reserving the 190px
height.

The right solution should be to display an empty graph, but I didn't
have the time to check how to do that.