fix: Show toggle as on after first enabling geolocation tracking
On first enabling geolocation tracking, the toggle is not set as on
because there is a bug in cozy-flagship-app. await
setGeolocationTracking(true)  return BEFORE the tracking is enabled.
The fix has been done in cozy-flagship-app
(bae397dd9060bb6122bfcdd2a1dab4e5cbb53687) but we need to wait for
the next release. So we fix it now for the moment by always
considering it as a success. It should be safe because starting
tracking should not fail.