feat: Add new DACC measures
Those measures are meant for an experimentation with Centre Val de
Loire.
It aggregates values such as CO2 or distance for the current month, and
group them based on the mode, school (when relevant), and user type
(when relevant).
Note the school matching relies on a pre-defined contacts group name,
including contacts' schools with geo information.

fix: Relationships