feat: Compute geojson aggregation in fetch service
We used to rely on the `timeseriesWithoutAggregation` service in order
to compute the `aggregation` block. This service is run after a geojson
creation, with 1min debounce.
This is not ideal because it is an async operation, which mean that some
timeseries can exist without the `aggregation`, forcing the app to adapt
queries and checks to guarantee its existence.

As the trips are now fetched through a coachco2 service (and not from an
external konnector), we can now compute this `aggregation` before saving
the trips.
This users will see their trips faster (better UX), and eventually we
won't have to check for this block existence anymore (better DX).