fix: The export must also have the modes chosen manually
In a past commit we removed the condition on
manual mode in the `getFeatureMode` function
(see c2c2ad0).

Today we need this condition because `getFeatureMode`
is also used by the Export function.

Without this condition, the latter exports the trips
without looking at the changes made by the user in the meantime.