feat: Request openpath purge after fetch
When the fetch service has saved new trips, request the openpath server
to purge itself from old data.

This is useful to prevent out-of-control database growing, and improve
pipeline performances, as we measured a performance drop when the
`Stage_analysis_timeseries` grows.

Note this purge is called only when new trips are saved, and is desgined
to keep a 7-days security before purging.