chore: Fix duplication of `createRoot` when reloading
For a long time we have had problems with the app's
routes opening modals.
Indeed, in dev we could not refresh the app with an
open modal without it being duplicated.

It is the fact of having the `createRoot` function
in the `init` function of the app that causes these problems,
because the latter executes twice when the app is launched
via the `yarn start/watch` commands.

Other Cozy apps (in the same versions of React)
execute this function in `setupApp` which is memoized,
and thus avoids the double execution of `createRoot`.