feat: Add time similarity for loop trips
We rely on similar start and end places to find recurring trips.
However, we cannot do this with "loop trips", i.e. with same start and
end place.
We used to rely on the distance for such trips, and group with similar
distance. But this is fragile, so we introduce a time metric
: a trip must now have a similar start time and similar
end time to be included. By "time", we mean hour, minutes, seconds.

So a trip the 01/03 from 12:00 to 13:00 will match a trip occuring the 02/03
from 11:50 to 13:10.

Note the distance filter is kept, but it is now less strict (distance ratio
increased from 0.1 to 0.3)